ScriptGeneration/requirements.txt
Flask
Flask-WTF
openai
python-dotenv
